+++ NEWS.md
@@ -1,3 +1,7 @@
+# tarflow.iquizoo 2.5.3
+
+* Added more tests. No user-level updates.
+
 # tarflow.iquizoo 2.5.2
 
 ## Bug Fixes